test(dataset): split the 2203-line test_dataset.py by concern (Theme E.1)
Break the 29-class + 2-function integration file into five focused modules —
test_dataset_{create,nodata,io,spatial,vectorize}.py — via the same AST splitter (bodies
extracted by line span, decorators preserved, per-file isort-sectioned imports). Pure move:
145 tests collected before and after, all green. Hoist the thrice-repeated 319-char WGS84
WKT literal into a wrapped WGS84_WKT constant and reflow one 121-char docstring so the new
files stay <=120. No test logic changes.
